move service to config
